feat(chartcuterie): Refactor Endpoint Chart to separate data
